// models/ekyc.model.ts
export interface EkycFormProps {
  displayName: string;
  fullName: string;
  idType?: number;
  idNo: string;
  laserId: string;
  docAddress: string;
  birthDate?: number;
  phoneNumber: string;
  ethnicitySt: string;
  ethnicity?: number;
  nationalitySt: string;
  nationality?: number;
  lineId: string;
  facebook: string;
  referralCode: string;
}